+++ bika/lims/utils/analysisrequest.py
@@ -300,6 +300,9 @@ def create_retest(ar):
     # Copy the analyses from the source
     intermediate_states = ['retracted', 'reflexed']
     for an in ar.getAnalyses(full_objects=True):
+        # skip retests
+        if an.isRetest():
+            continue
         if (api.get_workflow_status_of(an) in intermediate_states):
             # Exclude intermediate analyses
             continue
@@ -416,10 +419,9 @@ def create_partition(analysis_request, request, analyses, sample_type=None,
     # here because the transition "initialize" of analyses rely on a guard,
     # so the initialization can only be performed when the sample has been
     # received (DateReceived is set)
-    ActionHandlerPool.get_instance().queue_pool()
     for analysis in partition.getAnalyses(full_objects=True):
         doActionFor(analysis, "initialize")
-    ActionHandlerPool.get_instance().resume()
+        analysis.reindexObject()
     return partition